Comments on __arith__ and __metic__
It's all stable, but I needed to add at least SOME type of comment to what's going on so that I have something to review.

I also fixed the explanation of the processes in clq.md, and made the UndefinedError
in __metic__ offer more direct information about the failure. It's not perfect, but what permutative nonsense ever is?