MD5 implementation with Td
Instead of using the bytes type, the Tridozenal class handles everything. This approach is flawed in a few ways, but can still be used to accurately perform the checksum.

This was a fun exercise, even though I was close to tearing my hair out before I figured out the missing reversals. Still, I was happy to discover that the flaw was at no point due to cxr.base36.